fixes for mono

+++ src/Control/Monad/Supply.hs
@@ -12,7 +12,6 @@ import Control.Monad.Reader (MonadReader, MonadTrans)
 import Control.Monad (MonadPlus)
 import Control.Monad.State (StateT(..), MonadState(..))
 import Control.Monad.Writer (MonadWriter)
-import Data.Int (Int64)
 
 import Data.Functor.Identity (Identity(..))
 import Control.Monad.IO.Class (MonadIO)
@@ -21,15 +20,15 @@ import Control.Monad.Trans (lift)
 newtype SupplyT m a = SupplyT { unSupplyT :: StateT Integer m a }
   deriving (Functor, Applicative, Monad, MonadTrans, MonadError e, MonadWriter w, MonadReader r, Alternative, MonadPlus, MonadIO)
 
-runSupplyT :: Int64 -> SupplyT m a -> m (a, Int64)
+runSupplyT :: Integer -> SupplyT m a -> m (a, Integer)
 runSupplyT n = flip runStateT n . unSupplyT
 
-evalSupplyT :: (Functor m) => Int64 -> SupplyT m a -> m a
+evalSupplyT :: (Functor m) => Integer -> SupplyT m a -> m a
 evalSupplyT n = fmap fst . runSupplyT n
 
 type Supply = SupplyT Identity
 
-runSupply :: Int64 -> Supply a -> (a, Int64)
+runSupply :: Integer -> Supply a -> (a, Integer)
 runSupply n = runIdentity . runSupplyT n
 
 instance MonadState s m => MonadState s (SupplyT m) where
